RunManager optimizations
RunManager automatically generates PHSP events if no PHSP sample is set.
In that way it is possible to set a sample (e.g. a accepted phsp events
for unbinned efficiency correction) and use it for generation. In turn
for large sets it is not necessary to generater and save a hughe toy
sample.
Important bug fix in DataPoint