Cmake: Add linking option SAFE_LINKING
If enabled via "-DSAFE_LINKING=ON" ComPWA is linked statically
against libstdc++ and also boost is linked statically. Furthermore,
RPATH is set instead of RUNPATH. Note that libraries in
LIBRARY_PATH are not added to RPATH. Probably use `unset LIBRARY_PATH`.
There is still an issue with TBB. Unfortunatly TBB does not provide an
install target and libraries are not copied to CMAKE_INSTALL_PREFIX.
This needs to be fixed.