Fixed Interface to Geneva library (#223)
The interface code to the Geneva optimization library was previously not
adapted to the newest code, since it is not part of the continuous
integration. The code was refactored and improved. Also now the test
application runs successfully, so that the Geneva library can actually be
used. However be aware that the performance is currently not good. It is
unclear if this is intrinsic to Geneva or a bad configuration.
Some possible Geneva user interface improvements are mentioned in the
code.

- Rearranged optimizer tests
- Closes #221
- Geneva interface code now compiles.
- Geneva test are working (find correct minima). However the performance
  is extremely poor. Not sure what causes this.
- Adapted GenevaResult. However error estimates, covariance matrix and
  further properties are missing.
- The user can now specify an order of algorithms in the GenevaIF. These
  algorithms will be used in the minimization in exactly this order.
- Added example geneva steering config files