Refactored FourMomentum and Particle classes
The FourMomentum class was reworked, because the constructor is not
safe. In the RootDataIO test a FourMomentum with just a CMS energy was
created, but this created a wrong FourMomentum (px=CMS, e=0). This
problem is not eliminated. Also the FourMomentum and Particle classes
were a strange combination of OO and functional code.

* Improved safety and code quality of FourMomentum and Particle
constructors
* Improved code quality of this two class (removed weird functional code
parts)
* Removed unnecessary methods