Adapted other code to FunctionTree refactoring
These are mainly changes arising from the new FunctionTree construction
interface. The names for Intensities and Amplitudes were removed, since
they are not needed anymore. However to be able to calculated fit
fractions a `Component` xml tag was created, that can be assigned with a
unique label. These Components can then later be referenced to
calculate fit fractions.