refactor!: improve form factor code design (#293)
* Define form factor interface and implement existing form factors
accordingly. Dynamic functions like the BreitWigner receive form factor
interfaces. Closes #205.
* Add new standard relativistic BreitWigner description. It follows the
common definition and w.r.t to the current ComPWA defintion does not
include any analytic continuation (as this is not required for a BW
anyway). Closes #266.
* Fix issue of FormFactorStrategy being evaluated incorrectly.
Fixes #278.
* Adjust two tests to recent form factor change

Note: The IntegrationTest reference integral changed due the fixing of
the bug in the FormFactorStrategy and using a different BW model. This
was carefully verified and hence the reference value was adjusted to a
new correct value.

BREAKING CHANGE: Form factor type names change from integer strings to
labels "BlattWeisskopf" or "CrystalBarrel". Also the label for the
dynamic functions change. The "relativisticBreitWigner" label will
construct the new standard BW instead.