refactor!: move PIDs from Event to EventList (#306)
In the Event class, Particle has been replaced by FourMomentum. The particle information is contained within a list of PIDs that is contained within the EventList class. This required a lot of refactoring, but allows for checking the PIDs when for instance doing kinematic variable conversion.

Note that the format of RootDataIO has been changed as well: no use is made of the infamous TClonesArray anymore (note that even ROOT writes "It is recommended to use STL containers instead of TClonesArrays").

Also renamed several variables to make them agree with our naming conventions.


* feat: add interface getFinalState for Kinematics

This method is offered to allow the user to give an interpretation to a
data file of momentum tuples when using the DataIO functions.
--> Preparation for removal of Particle object

* refactor!: change Event structure to four-momenta

Define an Event as a list of four-momenta and a weight, instead of a
list of particles and weight. The particle information of the Event is a
constant across all Events belonging to this specific reaction.
Therefore extract this info and define it in a header-like part of a
EventList.

* Change Event structure and introduce EventList structure
* Remove Particle class and rename files
* Change data and kinematics interfaces accordingly
* Remove DataCorrection functionality, as this is a special part of data
  prepartion and does not belong into ComPWA

* feat: implement EventList in Kinematics

* refactor!: change generator interface

Change Generator interface to return EventCollections instead of single
Events.

* fix: add exception if no. of PIDs not same as 4mom

* refactor: use EventCollections in IntensityBuilder

Co-authored-by: Stefan Pflueger <spfluege@gmail.com>